backend/routes/image_routes.py
from fastapi import APIRouter, File, UploadFile, HTTPException
from fastapi.responses import JSONResponse
from PIL import Image
import io
import tensorflow as tf
import numpy as np

router = APIRouter()

# Load a pre-trained TensorFlow model for image analysis (e.g., object detection)
model = tf.saved_model.load("path_to_pretrained_model")

@router.post("/upload-image/")
async def upload_image(file: UploadFile = File(...)):
    try:
        # Read the image file
        image_data = await file.read()
        image = Image.open(io.BytesIO(image_data))

        # Convert image to numpy array
        image_array = np.array(image)

        # Perform image analysis (e.g., object detection)
        input_tensor = tf.convert_to_tensor(image_array)
        input_tensor = input_tensor[tf.newaxis, ...]
        detections = model(input_tensor)

        # Process detections (e.g., extract objects, labels, etc.)
        detected_objects = process_detections(detections)

        return JSONResponse(content={"message": "Image uploaded and analyzed", "detected_objects": detected_objects})
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

def process_detections(detections):
    # Dummy implementation for processing detections
    return ["object1", "object2"]
