backend/requirements.txt
# FastAPI & ASGI Server
fastapi==0.109.0
uvicorn==0.27.0

# Database
neo4j==5.14.0
sqlalchemy==2.0.23
python-multipart==0.0.6

# AI/ML Models
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
scikit-learn==1.3.2
joblib==1.3.2
onnxruntime==1.16.3
tensorflow==2.15.0    # For image analysis
Pillow==10.1.0         # For image processing

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
cryptography==41.0.5

# Utilities
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.1
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
httpx==0.25.2
